Fix array capacity calculation methods to use long values

+++ javatools_jitbridge/src/main/java/org/xtclang/ecstasy/collections/ArrayᐸCharᐳ.java
@@ -161,7 +161,7 @@ public class ArrayᐸCharᐳ
     }
 
     @Override
-    protected int $cap2len(int cap) {
+    protected long $cap2len(long cap) {
         return $utf21 ? $cap2len21bits(cap) : $cap2len8bits(cap);
     }
 }
