+++ javatools_jitbridge/src/main/java/org/xtclang/ecstasy/collections/ArrayᐸDec32ᐳ.java
@@ -127,7 +127,7 @@ public class ArrayᐸDec32ᐳ
     }
 
     @Override
-    protected int $cap2len(int cap) {
+    protected long $cap2len(long cap) {
         return $cap2len32bits(cap);
     }
 }
