+++ javatools_jitbridge/src/main/java/org/xtclang/ecstasy/collections/ArrayᐸDec64ᐳ.java
@@ -130,7 +130,7 @@ public class ArrayᐸDec64ᐳ
     }
 
     @Override
-    protected int $cap2len(int cap) {
+    protected long $cap2len(long cap) {
         return cap;
     }
 }
