+++ javatools_jitbridge/src/main/java/org/xtclang/ecstasy/collections/ArrayᐸFloat64ᐳ.java
@@ -131,7 +131,7 @@ public class ArrayᐸFloat64ᐳ
     }
 
     @Override
-    protected int $cap2len(int cap) {
+    protected long $cap2len(long cap) {
         return cap;
     }
 }
