+++ javatools_jitbridge/src/main/java/org/xtclang/ecstasy/collections/ArrayᐸUInt32ᐳ.java
@@ -131,7 +131,7 @@ public class ArrayᐸUInt32ᐳ
     }
 
     @Override
-    protected int $cap2len(int cap) {
+    protected long $cap2len(long cap) {
         return $cap2len32bits(cap);
     }
 }
